Fixes part of oppia#10415: Refactor story validators - modified
