+++ core/domain/prod_validators.py
@@ -43,8 +43,6 @@ from core.domain import rights_domain
 from core.domain import rights_manager
 from core.domain import skill_fetchers
 from core.domain import stats_services
-from core.domain import story_domain
-from core.domain import story_fetchers
 from core.domain import subtopic_page_domain
 from core.domain import subtopic_page_services
 from core.domain import suggestion_services
@@ -2254,191 +2252,6 @@ class TopicSimilaritiesModelValidator(base_model_validators.BaseModelValidator):
         return [cls._validate_topic_similarities]
 
 
-class StoryModelValidator(base_model_validators.BaseModelValidator):
-    """Class for validating StoryModel."""
-
-    @classmethod
-    def _get_model_domain_object_instance(cls, item):
-        return story_fetchers.get_story_from_model(item)
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        snapshot_model_ids = [
-            '%s-%d' % (item.id, version)
-            for version in python_utils.RANGE(1, item.version + 1)]
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'story_commit_log_entry_ids',
-                story_models.StoryCommitLogEntryModel,
-                ['story-%s-%s'
-                 % (item.id, version) for version in python_utils.RANGE(
-                     1, item.version + 1)]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'story_summary_ids',
-                story_models.StorySummaryModel, [item.id]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_metadata_ids',
-                story_models.StorySnapshotMetadataModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'snapshot_content_ids',
-                story_models.StorySnapshotContentModel,
-                snapshot_model_ids),
-            base_model_validators.ExternalModelFetcherDetails(
-                'exploration_ids',
-                exp_models.ExplorationModel,
-                [node['exploration_id'] for node in (
-                    item.story_contents['nodes'])])]
-
-
-class StorySnapshotMetadataModelValidator(
-        base_model_validators.BaseSnapshotMetadataModelValidator):
-    """Class for validating StorySnapshotMetadataModel."""
-
-    EXTERNAL_MODEL_NAME = 'story'
-
-    @classmethod
-    def _get_change_domain_class(cls, unused_item):
-        return story_domain.StoryChange
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'story_ids', story_models.StoryModel,
-                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'committer_ids', user_models.UserSettingsModel,
-                [item.committer_id])]
-
-
-class StorySnapshotContentModelValidator(
-        base_model_validators.BaseSnapshotContentModelValidator):
-    """Class for validating StorySnapshotContentModel."""
-
-    EXTERNAL_MODEL_NAME = 'story'
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'story_ids', story_models.StoryModel,
-                [item.id[:item.id.rfind(base_models.VERSION_DELIMITER)]])]
-
-
-class StoryCommitLogEntryModelValidator(
-        base_model_validators.BaseCommitLogEntryModelValidator):
-    """Class for validating StoryCommitLogEntryModel."""
-
-    EXTERNAL_MODEL_NAME = 'story'
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        # Valid id: [story]-[story_id]-[story_version].
-        regex_string = '^(story)-%s-\\d+$' % (
-            item.story_id)
-
-        return regex_string
-
-    @classmethod
-    def _get_change_domain_class(cls, item):
-        if item.id.startswith('story'):
-            return story_domain.StoryChange
-        else:
-            cls._add_error(
-                'model %s' % base_model_validators.ERROR_CATEGORY_ID_CHECK,
-                'Entity id %s: Entity id does not match regex pattern' % (
-                    item.id))
-            return None
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'story_ids', story_models.StoryModel, [item.story_id]),
-        ]
-
-
-class StorySummaryModelValidator(
-        base_model_validators.BaseSummaryModelValidator):
-    """Class for validating StorySummaryModel."""
-
-    @classmethod
-    def _get_model_domain_object_instance(cls, item):
-        return story_fetchers.get_story_summary_from_model(item)
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'story_ids', story_models.StoryModel, [item.id])]
-
-    @classmethod
-    def _validate_node_titles(
-            cls, item, field_name_to_external_model_references):
-        """Validate that node_titles of model is equal to list of node titles
-        in StoryModel.story_contents.
-
-        Args:
-            item: datastore_services.Model. StorySummaryModel to validate.
-            field_name_to_external_model_references:
-                dict(str, (list(base_model_validators.ExternalModelReference))).
-                A dict keyed by field name. The field name represents
-                a unique identifier provided by the storage
-                model to which the external model is associated. Each value
-                contains a list of ExternalModelReference objects corresponding
-                to the field_name. For examples, all the external Exploration
-                Models corresponding to a storage model can be associated
-                with the field name 'exp_ids'. This dict is used for
-                validation of External Model properties linked to the
-                storage model.
-        """
-        story_model_references = (
-            field_name_to_external_model_references['story_ids'])
-
-        for story_model_reference in story_model_references:
-            story_model = story_model_reference.model_instance
-            if story_model is None or story_model.deleted:
-                model_class = story_model_reference.model_class
-                model_id = story_model_reference.model_id
-                cls._add_error(
-                    'story_ids %s' % (
-                        base_model_validators.ERROR_CATEGORY_FIELD_CHECK),
-                    'Entity id %s: based on field story_ids having'
-                    ' value %s, expected model %s with id %s but it doesn\'t'
-                    ' exist' % (
-                        item.id, model_id, model_class.__name__, model_id))
-                continue
-            nodes = story_model.story_contents['nodes']
-            node_titles = [node.title for node in nodes]
-            if item.node_titles != node_titles:
-                cls._add_error(
-                    'node titles check',
-                    'Entity id %s: Node titles: %s does not match the '
-                    'nodes in story_contents dict: %s' % (
-                        item.id, item.node_titles, nodes))
-
-    @classmethod
-    def _get_external_model_properties(cls):
-        story_model_properties_dict = {
-            'title': 'title',
-            'language_code': 'language_code',
-            'description': 'description',
-            'story_model_created_on': 'created_on',
-            'story_model_last_updated': 'last_updated'
-        }
-
-        return [(
-            'story',
-            'story_ids',
-            story_model_properties_dict
-        )]
-
-    @classmethod
-    def _get_external_instance_custom_validation_functions(cls):
-        return [cls._validate_node_titles]
-
-
 class GeneralSuggestionModelValidator(base_model_validators.BaseModelValidator):
     """Class for validating GeneralSuggestionModels."""
 
@@ -4532,173 +4345,6 @@ class CollectionProgressModelValidator(
             cls._validate_collections_are_public,
             cls._validate_completed_exploration
         ]
-
-
-class StoryProgressModelValidator(base_model_validators.BaseUserModelValidator):
-    """Class for validating StoryProgressModels."""
-
-    @classmethod
-    def _get_model_id_regex(cls, item):
-        return '^%s\\.%s$' % (item.user_id, item.story_id)
-
-    @classmethod
-    def _get_external_id_relationships(cls, item):
-        return [
-            base_model_validators.ExternalModelFetcherDetails(
-                'user_settings_ids', user_models.UserSettingsModel,
-                [item.user_id]),
-            base_model_validators.ExternalModelFetcherDetails(
-                'story_ids', story_models.StoryModel, [item.story_id])]
-
-    @classmethod
-    def _validate_story_is_public(
-            cls, item, field_name_to_external_model_references):
-        """Validates that story is public.
-
-        Args:
-            item: datastore_services.Model. StoryProgressModel to validate.
-            field_name_to_external_model_references:
-                dict(str, (list(base_model_validators.ExternalModelReference))).
-                A dict keyed by field name. The field name represents
-                a unique identifier provided by the storage
-                model to which the external model is associated. Each value
-                contains a list of ExternalModelReference objects corresponding
-                to the field_name. For examples, all the external Exploration
-                Models corresponding to a storage model can be associated
-                with the field name 'exp_ids'. This dict is used for
-                validation of External Model properties linked to the
-                storage model.
-        """
-        story_model_references = (
-            field_name_to_external_model_references['story_ids'])
-
-        for story_model_reference in story_model_references:
-            story_model = story_model_reference.model_instance
-            if story_model is None or story_model.deleted:
-                model_class = story_model_reference.model_class
-                model_id = story_model_reference.model_id
-                cls._add_error(
-                    'story_ids %s' % (
-                        base_model_validators.ERROR_CATEGORY_FIELD_CHECK),
-                    'Entity id %s: based on field story_ids having'
-                    ' value %s, expected model %s with id %s but it doesn\'t'
-                    ' exist' % (
-                        item.id, model_id, model_class.__name__, model_id))
-                continue
-            topic_id = story_model.corresponding_topic_id
-            if topic_id:
-                topic = topic_models.TopicModel.get_by_id(topic_id)
-                all_story_references = (
-                    topic.canonical_story_references +
-                    topic.additional_story_references)
-                story_is_published = False
-                for reference in all_story_references:
-                    if reference['story_id'] == story_model.id:
-                        story_is_published = reference['story_is_published']
-                if not story_is_published:
-                    cls._add_error(
-                        'public story check',
-                        'Entity id %s: Story with id %s corresponding to '
-                        'entity is private' % (item.id, story_model.id))
-
-    @classmethod
-    def _validate_completed_nodes(
-            cls, item, field_name_to_external_model_references):
-        """Validates that completed nodes belong to the story.
-
-        Args:
-            item: datastore_services.Model. StoryProgressModel to validate.
-            field_name_to_external_model_references:
-                dict(str, (list(base_model_validators.ExternalModelReference))).
-                A dict keyed by field name. The field name represents
-                a unique identifier provided by the storage
-                model to which the external model is associated. Each value
-                contains a list of ExternalModelReference objects corresponding
-                to the field_name. For examples, all the external Exploration
-                Models corresponding to a storage model can be associated
-                with the field name 'exp_ids'. This dict is used for
-                validation of External Model properties linked to the
-                storage model.
-        """
-        completed_activity_model = user_models.CompletedActivitiesModel.get(
-            item.user_id)
-        story_model_references = (
-            field_name_to_external_model_references['story_ids'])
-
-        for story_model_reference in story_model_references:
-            story_model = story_model_reference.model_instance
-            if story_model is None or story_model.deleted:
-                model_class = story_model_reference.model_class
-                model_id = story_model_reference.model_id
-                cls._add_error(
-                    'story_ids %s' % (
-                        base_model_validators.ERROR_CATEGORY_FIELD_CHECK),
-                    'Entity id %s: based on field story_ids having'
-                    ' value %s, expected model %s with id %s but it doesn\'t'
-                    ' exist' % (
-                        item.id, model_id, model_class.__name__, model_id))
-                continue
-            story_node_ids = [
-                node['id'] for node in story_model.story_contents['nodes']]
-            invalid_node_ids = [
-                node_id
-                for node_id in item.completed_node_ids if node_id not in (
-                    story_node_ids)]
-            if invalid_node_ids:
-                cls._add_error(
-                    'completed node check',
-                    'Entity id %s: Following completed node ids %s do '
-                    'not belong to the story with id %s corresponding '
-                    'to the entity' % (
-                        item.id, invalid_node_ids, story_model.id))
-
-            # Checks that the explorations corresponding to completed nodes
-            # exist, are marked as completed in CompletedActivitiesModel
-            # and are public.
-            private_exp_ids = []
-            missing_exp_ids = []
-            unmarked_exp_ids = []
-            completed_exp_ids = []
-            for node in story_model.story_contents['nodes']:
-                if node['id'] in item.completed_node_ids:
-                    completed_exp_ids.append(node['exploration_id'])
-                    if node['exploration_id'] not in (
-                            completed_activity_model.exploration_ids):
-                        unmarked_exp_ids.append(node['exploration_id'])
-                    if rights_manager.is_exploration_private(
-                            node['exploration_id']):
-                        private_exp_ids.append(node['exploration_id'])
-
-            exp_model_list = exp_models.ExplorationModel.get_multi(
-                completed_exp_ids)
-            for index, exp_model in enumerate(exp_model_list):
-                if exp_model is None or exp_model.deleted:
-                    missing_exp_ids.append(completed_exp_ids[index])
-
-            error_msg = ''
-            if private_exp_ids:
-                error_msg = error_msg + (
-                    'Following exploration ids are private %s. ' % (
-                        private_exp_ids))
-            if missing_exp_ids:
-                error_msg = error_msg + (
-                    'Following exploration ids are missing %s. ' % (
-                        missing_exp_ids))
-            if unmarked_exp_ids:
-                error_msg = error_msg + (
-                    'Following exploration ids are not marked in '
-                    'CompletedActivitiesModel %s.' % unmarked_exp_ids)
-
-            if error_msg:
-                cls._add_error(
-                    'explorations in completed node check',
-                    'Entity id %s: %s' % (item.id, error_msg))
-
-    @classmethod
-    def _get_external_instance_custom_validation_functions(cls):
-        return [
-            cls._validate_story_is_public,
-            cls._validate_completed_nodes]
 
 
 class UserQueryModelValidator(base_model_validators.BaseUserModelValidator):
